backend/requirements.txt
Flask>=3.1.3
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.1.0
Flask-CORS> 6.0.2
python-dotenv>=1.2.1
psycopg2-binary>=2.9.9
